Extract tst_Map_Initialization
