+++ lib/TbMdlLib/test/src/tst_Map.cpp
@@ -18,7 +18,6 @@
  */
 
 #include "Logger.h"
-#include "Observer.h"
 #include "TestLogger.h"
 #include "fs/TestEnvironment.h"
 #include "gl/Material.h"
@@ -52,14 +51,11 @@
 #include "mdl/TagMatcher.h"
 #include "mdl/TestFactory.h"
 #include "mdl/TestUtils.h"
-#include "mdl/TransactionScope.h"
 #include "mdl/UpdateBrushFaceAttributes.h"
 #include "mdl/WorldNode.h"
 
-#include "kd/k.h"
 #include "kd/vector_utils.h"
 
-#include "vm/approx.h"
 #include "vm/vec_io.h" // IWYU pragma: keep
 
 #include <catch2/catch_test_macros.hpp>
@@ -88,397 +84,11 @@ public:
   size_t selectOption(const std::vector<std::string>&) override { return m_option; }
 };
 
-auto makeAbsolute(const auto& path)
-{
-  return std::filesystem::current_path() / path;
-}
-
 } // namespace
 
 TEST_CASE("Map")
 {
   auto environmentConfig = EnvironmentConfig{};
-
-  SECTION("create")
-  {
-    auto taskManager = createTestTaskManager();
-    auto resourceManager = gl::ResourceManager{};
-    auto logger = NullLogger{};
-
-    SECTION("Calling create sets worldspawn and notifies observers")
-    {
-      Map::createMap(
-        environmentConfig,
-        DefaultGameInfo,
-        DefaultGameInfo.gamePathPreference.defaultValue,
-        MapFormat::Standard,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            CHECK(
-              map->worldNode().entity()
-              == Entity{{
-                {EntityPropertyKeys::Classname,
-                 EntityPropertyValues::WorldspawnClassname},
-              }});
-            CHECK(!map->worldBounds().is_empty());
-            CHECK(!map->persistent());
-            CHECK(map->path() == "unnamed.map");
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
-
-    SECTION("Loads the initial map if configured")
-    {
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.forceEmptyNewMap = false;
-      gameInfo.gameConfig.path = "fixture/test/mdl/Map/GameConfig.cfg";
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Valve", {"initialMap.map"}},
-      };
-
-      Map::createMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Valve,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            const auto* defaultLayerNode = map->worldNode().defaultLayer();
-            REQUIRE(defaultLayerNode->children().size() == 1);
-
-            const auto* brushNode =
-              dynamic_cast<const BrushNode*>(defaultLayerNode->children().front());
-            REQUIRE(brushNode);
-
-            CHECK(
-              brushNode->brush().bounds() == vm::bbox3d{{-32, -32, -64}, {32, 32, 64}});
-
-            const auto* valveVersionProperty =
-              map->worldNode().entity().property(EntityPropertyKeys::ValveVersion);
-            REQUIRE(valveVersionProperty);
-            CHECK(*valveVersionProperty == "220");
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
-
-    SECTION("Sets Valve version property")
-    {
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.forceEmptyNewMap = false;
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Valve", {"initialMap.map"}},
-      };
-
-      Map::createMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Valve,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            const auto* valveVersionProperty =
-              map->worldNode().entity().property(EntityPropertyKeys::ValveVersion);
-            REQUIRE(valveVersionProperty);
-            CHECK(*valveVersionProperty == "220");
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
-
-    SECTION("Sets material config property")
-    {
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.forceEmptyNewMap = false;
-      gameInfo.gameConfig.materialConfig.property = "wad";
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Valve", {"initialMap.map"}},
-      };
-
-      Map::createMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Valve,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            const auto* materialConfigProperty =
-              map->worldNode().entity().property("wad");
-            REQUIRE(materialConfigProperty);
-            CHECK(*materialConfigProperty == "");
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
-
-    SECTION("Creates an initial brush if configured")
-    {
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.forceEmptyNewMap = false;
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Standard", {}},
-      };
-
-      Map::createMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Valve,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            const auto* defaultLayerNode = map->worldNode().defaultLayer();
-            REQUIRE(defaultLayerNode->children().size() == 1);
-
-            const auto* brushNode =
-              dynamic_cast<const BrushNode*>(defaultLayerNode->children().front());
-            REQUIRE(brushNode);
-
-            CHECK(
-              brushNode->brush().bounds() == vm::bbox3d{{-64, -64, -16}, {64, 64, 16}});
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
-
-    SECTION("Loads default entity definition file")
-    {
-      auto env = fs::TestEnvironment{};
-      env.createFile("Quake.fgd", R"(@SolidClass = some_entity : "Some Entity" [])");
-
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.path = env.dir() / "GameConfig.cfg";
-      gameInfo.gameConfig.entityConfig.defFilePaths.emplace_back("Quake.fgd");
-
-
-      Map::createMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Standard,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            REQUIRE(map->entityDefinitionManager().definitions().size() == 1);
-            CHECK(
-              map->entityDefinitionManager().definitions().front().name == "some_entity");
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
-  }
-
-  SECTION("loadMap")
-  {
-    auto taskManager = createTestTaskManager();
-    auto resourceManager = gl::ResourceManager{};
-    auto logger = NullLogger{};
-
-    SECTION("Sets world bounds, game and file path")
-    {
-      const auto worldBounds = vm::bbox3d{8192.0};
-      const auto path = makeAbsolute("fixture/test/mdl/Map/emptyValveMap.map");
-
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Valve", {}},
-      };
-
-      Map::loadMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Unknown,
-        worldBounds,
-        path,
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([&](auto map) {
-            CHECK(map->worldBounds() == worldBounds);
-            CHECK(map->path() == path);
-            CHECK(map->persistent());
-          })
-        | kdl::transform_error([](auto e) { FAIL(e.msg); });
-    }
-
-    SECTION("Format detection")
-    {
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Valve", {}},
-        {"Standard", {}},
-        {"Quake3", {}},
-      };
-
-      SECTION("Detect Valve Format Map")
-      {
-        Map::loadMap(
-          environmentConfig,
-          gameInfo,
-          gameInfo.gamePathPreference.defaultValue,
-          MapFormat::Unknown,
-          vm::bbox3d{8192.0},
-          makeAbsolute("fixture/test/mdl/Map/valveFormatMapWithoutFormatTag.map"),
-          *taskManager,
-          resourceManager,
-          logger)
-          | kdl::transform([&](auto map) {
-              CHECK(map->worldNode().mapFormat() == mdl::MapFormat::Valve);
-              CHECK(map->worldNode().defaultLayer()->childCount() == 1);
-            })
-          | kdl::transform_error([](auto e) { FAIL(e.msg); });
-      }
-
-      SECTION("Detect Standard Format Map")
-      {
-        Map::loadMap(
-          environmentConfig,
-          gameInfo,
-          gameInfo.gamePathPreference.defaultValue,
-          MapFormat::Unknown,
-          vm::bbox3d{8192.0},
-          makeAbsolute("fixture/test/mdl/Map/standardFormatMapWithoutFormatTag.map"),
-          *taskManager,
-          resourceManager,
-          logger)
-          | kdl::transform([&](auto map) {
-              CHECK(map->worldNode().mapFormat() == mdl::MapFormat::Standard);
-              CHECK(map->worldNode().defaultLayer()->childCount() == 1);
-            })
-          | kdl::transform_error([](auto e) { FAIL(e.msg); });
-      }
-
-      SECTION("detectEmptyMap")
-      {
-        Map::loadMap(
-          environmentConfig,
-          gameInfo,
-          gameInfo.gamePathPreference.defaultValue,
-          MapFormat::Unknown,
-          vm::bbox3d{8192.0},
-          makeAbsolute("fixture/test/mdl/Map/emptyMapWithoutFormatTag.map"),
-          *taskManager,
-          resourceManager,
-          logger)
-          | kdl::transform([&](auto map) {
-              // an empty map detects as Valve because Valve is listed first in the game
-              // config
-              CHECK(map->worldNode().mapFormat() == mdl::MapFormat::Valve);
-              CHECK(map->worldNode().defaultLayer()->childCount() == 0);
-            })
-          | kdl::transform_error([](auto e) { FAIL(e.msg); });
-      }
-
-      SECTION("mixedFormats")
-      {
-        // map has both Standard and Valve brushes
-        CHECK(!Map::loadMap(
-          environmentConfig,
-          gameInfo,
-          gameInfo.gamePathPreference.defaultValue,
-          MapFormat::Unknown,
-          vm::bbox3d{8192.0},
-          makeAbsolute("fixture/test/mdl/Map/mixedFormats.map"),
-          *taskManager,
-          resourceManager,
-          logger));
-      }
-    }
-
-    SECTION("Loads default entity definition file")
-    {
-      auto env = fs::TestEnvironment{};
-      env.createFile("Quake.fgd", R"(@SolidClass = some_entity : "Some Entity" [])");
-
-      auto gameInfo = DefaultGameInfo;
-      gameInfo.gameConfig.path = env.dir() / "GameConfig.cfg";
-      gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-        {"Valve", {}},
-      };
-      gameInfo.gameConfig.entityConfig.defFilePaths.emplace_back("Quake.fgd");
-
-
-      Map::loadMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Unknown,
-        vm::bbox3d{8192.0},
-        makeAbsolute("fixture/test/mdl/Map/valveFormatMapWithoutFormatTag.map"),
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([&](auto map) {
-            REQUIRE(map->entityDefinitionManager().definitions().size() == 1);
-            CHECK(
-              map->entityDefinitionManager().definitions().front().name == "some_entity");
-          })
-        | kdl::transform_error([](auto e) { FAIL(e.msg); });
-    }
-  }
-
-  SECTION("reload")
-  {
-    auto taskManager = createTestTaskManager();
-    auto resourceManager = gl::ResourceManager{};
-    auto logger = NullLogger{};
-
-    auto gameInfo = DefaultGameInfo;
-    gameInfo.gameConfig.fileFormats = std::vector<MapFormatConfig>{
-      {"Valve", {}},
-    };
-
-    const auto worldBounds = vm::bbox3d{8192.0};
-    const auto path = makeAbsolute("fixture/test/mdl/Map/emptyValveMap.map");
-
-    Map::loadMap(
-      environmentConfig,
-      gameInfo,
-      gameInfo.gamePathPreference.defaultValue,
-      MapFormat::Unknown,
-      vm::bbox3d{8192.0},
-      path,
-      *taskManager,
-      resourceManager,
-      logger)
-      | kdl::and_then([&](auto map) {
-          REQUIRE(map->worldBounds() == worldBounds);
-          REQUIRE(&map->gameInfo() == &gameInfo);
-          REQUIRE(map->path() == path);
-          REQUIRE(map->persistent());
-
-          auto* transientEntityNode = new EntityNode{{}};
-          addNodes(*map, {{parentForNodes(*map), {transientEntityNode}}});
-          REQUIRE(
-            map->worldNode().defaultLayer()->children()
-            == std::vector<Node*>{transientEntityNode});
-          REQUIRE(map->modified());
-
-          return map->reload() | kdl::transform([&](auto reloadedMap) {
-                   CHECK(reloadedMap->worldBounds() == worldBounds);
-                   CHECK(&reloadedMap->gameInfo() == &gameInfo);
-                   CHECK(reloadedMap->path() == path);
-                   CHECK(reloadedMap->persistent());
-                   CHECK(!reloadedMap->modified());
-                   CHECK(
-                     reloadedMap->worldNode().defaultLayer()->children()
-                     == std::vector<Node*>{});
-                 });
-        })
-      | kdl::transform_error([](auto e) { FAIL(e.msg); });
-  }
 
   SECTION("persistent")
   {
@@ -1503,70 +1113,6 @@ TEST_CASE("Map")
           CHECK(logger.countMessages(LogLevel::Warn) > warnCountBefore);
         })
       | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-  }
-
-  SECTION("Entity definition file handling")
-  {
-    auto taskManager = createTestTaskManager();
-    auto resourceManager = gl::ResourceManager{};
-    auto logger = NullLogger{};
-
-    SECTION("Add and convert properties")
-    {
-      using namespace EntityPropertyKeys;
-
-      auto env = fs::TestEnvironment{};
-      env.createFile("Quake.fgd", R"(@PointClass = some_entity : "Some Entity" [])");
-
-      auto gameInfo = QuakeGameInfo;
-      gameInfo.gameConfig.path = env.dir() / "GameConfig.cfg";
-      gameInfo.gameConfig.entityConfig.defFilePaths.emplace_back("Quake.fgd");
-
-
-      Map::createMap(
-        environmentConfig,
-        gameInfo,
-        gameInfo.gamePathPreference.defaultValue,
-        MapFormat::Standard,
-        vm::bbox3d{8192.0},
-        *taskManager,
-        resourceManager,
-        logger)
-        | kdl::transform([](auto map) {
-            CHECK(
-              map->entityDefinitionManager().definitions()
-              == std::vector{EntityDefinition{
-                "some_entity",
-                RgbaF{0.6f, 0.6f, 0.6f, 1.0f},
-                "Some Entity",
-                {
-                  PropertyDefinition{
-                    Target,
-                    PropertyValueTypes::LinkSource{},
-                    "name of entity to trigger",
-                    "generated by TrenchBroom"},
-                  PropertyDefinition{
-                    Killtarget,
-                    PropertyValueTypes::LinkSource{},
-                    "name of entity to kill",
-                    "generated by TrenchBroom"},
-                  PropertyDefinition{
-                    Targetname,
-                    PropertyValueTypes::LinkTarget{},
-                    "target name for linking",
-                    "generated by TrenchBroom"},
-                  PropertyDefinition{
-                    Origin,
-                    PropertyValueTypes::Origin{},
-                    "point entity origin",
-                    "generated by TrenchBroom"},
-
-                },
-                PointEntityDefinition{vm::bbox3d{8.0}, {}, {}},
-                1}});
-          })
-        | kdl::transform_error([](const auto& e) { FAIL(e.msg); });
-    }
   }
 
   SECTION("Duplicate and Copy / Paste behave identically")
